move estimate_predicates test
